feat: configure grouped series
