+++ packages/frontend/src/components/ChartConfigPanel/ChartConfigTabs.tsx
@@ -1,14 +1,5 @@
+import { InputGroup, Tab, Tabs } from '@blueprintjs/core';
 import {
-    Button,
-    Collapse,
-    Colors,
-    HTMLSelect,
-    InputGroup,
-    Tab,
-    Tabs,
-} from '@blueprintjs/core';
-import {
-    Field,
     fieldId,
     getDefaultSeriesColor,
     getDimensions,
@@ -16,300 +7,21 @@ import {
     getItemLabel,
     getMetrics,
     getSeriesId,
-    isDimension,
     isField,
     Metric,
-    Series,
     TableCalculation,
 } from 'common';
 import React, { FC, useCallback, useMemo, useState } from 'react';
-import { useToggle } from 'react-use';
-import { getDimensionFormatter } from '../../utils/resultFormatter';
 import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
 import {
     FieldsGrid,
     GridLabel,
-    GroupSeriesBlock,
     InputWrapper,
-    SeriesBlock,
-    SeriesExtraInputs,
-    SeriesExtraInputWrapper,
-    SeriesMainInputs,
-    SeriesTitle,
-    SeriesWrapper,
     Wrapper,
 } from './ChartConfigPanel.styles';
 import FieldLayoutOptions from './FieldLayoutOptions';
-import SeriesColorPicker from './SeriesColorPicker';
-
-const getFormatterValue = (
-    value: any,
-    key: string,
-    items: Array<Field | TableCalculation>,
-) => {
-    const item = items.find((i) => getItemId(i) === key);
-    const fieldFormatter =
-        item && isField(item) && isDimension(item)
-            ? getDimensionFormatter(item)
-            : null;
-    return fieldFormatter?.({ value: value }) ?? `${value}`;
-};
-
-const ValueLabelsInput: FC<{
-    label?: React.ReactNode;
-    value?: Series['label'];
-    onLabelChange: (label: Series['label']) => void;
-}> = ({ label, value, onLabelChange }) => (
-    <SeriesExtraInputWrapper label={label || 'Value labels'}>
-        <HTMLSelect
-            fill
-            value={value?.position || 'hidden'}
-            options={[
-                { value: 'hidden', label: 'Hidden' },
-                { value: 'top', label: 'Top' },
-                { value: 'bottom', label: 'Bottom' },
-                { value: 'left', label: 'Left' },
-                { value: 'right', label: 'Right' },
-            ]}
-            onChange={(e) => {
-                const option = e.target.value;
-                onLabelChange(
-                    option === 'hidden'
-                        ? { show: false }
-                        : {
-                              show: true,
-                              position: option as any,
-                          },
-                );
-            }}
-        />
-    </SeriesExtraInputWrapper>
-);
-
-type SeriesConfigurationProps = {
-    isCollapsable?: boolean;
-    placeholderName: string;
-    series: Series;
-    fallbackColor?: string;
-    onColorChange: (color: string) => void;
-    onNameChange: (name: string | undefined) => void;
-    onLabelChange: (label: Series['label']) => void;
-};
-
-const SeriesConfiguration: FC<SeriesConfigurationProps> = ({
-    isCollapsable,
-    placeholderName,
-    series,
-    fallbackColor,
-    onColorChange,
-    onNameChange,
-    onLabelChange,
-}) => {
-    const [isOpen, toggleIsOpen] = useToggle(false);
-    return (
-        <SeriesWrapper>
-            <SeriesMainInputs>
-                <SeriesColorPicker
-                    color={series.color || fallbackColor}
-                    onChange={onColorChange}
-                />
-                <InputGroup
-                    fill
-                    placeholder={placeholderName}
-                    defaultValue={series.name}
-                    onBlur={(e) => onNameChange(e.currentTarget.value)}
-                />
-                {isCollapsable && (
-                    <Button
-                        icon={isOpen ? 'caret-up' : 'caret-down'}
-                        onClick={toggleIsOpen}
-                    />
-                )}
-            </SeriesMainInputs>
-            <Collapse isOpen={!isCollapsable || isOpen}>
-                <SeriesExtraInputs>
-                    <ValueLabelsInput
-                        value={series.label}
-                        onLabelChange={onLabelChange}
-                    />
-                </SeriesExtraInputs>
-            </Collapse>
-        </SeriesWrapper>
-    );
-};
-
-type GroupedSeriesConfigurationProps = {
-    groupedSeries: Record<string, Series[]>;
-    items: Array<Field | TableCalculation>;
-    getSeriesColor: (key: string) => string | undefined;
-    updateAllGroupedSeries: (fieldKey: string, series: Partial<Series>) => void;
-    updateSingleSeries: (series: Series) => void;
-};
-
-const GroupedSeriesConfiguration: FC<GroupedSeriesConfigurationProps> = ({
-    groupedSeries,
-    items,
-    getSeriesColor,
-    updateSingleSeries,
-    updateAllGroupedSeries,
-}) => {
-    return (
-        <>
-            {Object.entries(groupedSeries).map(([fieldKey, seriesGroup]) => {
-                const field = items.find(
-                    (item) => getItemId(item) === fieldKey,
-                );
-                if (!field) {
-                    return (
-                        <SeriesBlock>
-                            <span
-                                style={{
-                                    width: '100%',
-                                    color: Colors.GRAY1,
-                                }}
-                            >
-                                Tried to reference field with unknown id:{' '}
-                                {fieldKey}
-                            </span>
-                        </SeriesBlock>
-                    );
-                }
-                const isLabelTheSameForAllSeries =
-                    new Set(seriesGroup.map(({ label }) => label?.position))
-                        .size === 1;
-                return (
-                    <GroupSeriesBlock>
-                        <SeriesTitle>{getItemLabel(field)}</SeriesTitle>
-                        <SeriesExtraInputs>
-                            <ValueLabelsInput
-                                label={
-                                    !isLabelTheSameForAllSeries ? (
-                                        <span>
-                                            Value labels{' '}
-                                            <span
-                                                style={{ color: Colors.RED1 }}
-                                            >
-                                                (!)
-                                            </span>
-                                        </span>
-                                    ) : undefined
-                                }
-                                value={seriesGroup[0].label}
-                                onLabelChange={(label) =>
-                                    updateAllGroupedSeries(fieldKey, { label })
-                                }
-                            />
-                        </SeriesExtraInputs>
-                        {seriesGroup?.map((series) => {
-                            const formattedValue = getFormatterValue(
-                                series.encode.yRef.pivotValues![0].value,
-                                series.encode.yRef.pivotValues![0].field,
-                                items,
-                            );
-                            return (
-                                <SeriesConfiguration
-                                    isCollapsable
-                                    series={series}
-                                    placeholderName={`[${formattedValue}] ${getItemLabel(
-                                        field,
-                                    )}`}
-                                    fallbackColor={getSeriesColor(
-                                        getSeriesId(series),
-                                    )}
-                                    onColorChange={(color) => {
-                                        updateSingleSeries({
-                                            ...series,
-                                            color,
-                                        });
-                                    }}
-                                    onNameChange={(name) =>
-                                        updateSingleSeries({
-                                            ...series,
-                                            name,
-                                        })
-                                    }
-                                    onLabelChange={(label) => {
-                                        updateSingleSeries({
-                                            ...series,
-                                            label,
-                                        });
-                                    }}
-                                />
-                            );
-                        })}
-                    </GroupSeriesBlock>
-                );
-            })}
-        </>
-    );
-};
-
-type BasicSeriesConfigurationProps = {
-    series?: Series[];
-    items: Array<Field | TableCalculation>;
-    getSeriesColor: (key: string) => string | undefined;
-    updateSingleSeries: (series: Series) => void;
-};
-
-const BasicSeriesConfiguration: FC<BasicSeriesConfigurationProps> = ({
-    series: allSeries,
-    items,
-    getSeriesColor,
-    updateSingleSeries,
-}) => {
-    return (
-        <>
-            {allSeries?.map((series) => {
-                const field = items.find(
-                    (item) => getItemId(item) === series.encode.yRef.field,
-                );
-                if (!field) {
-                    return (
-                        <SeriesBlock>
-                            <span
-                                style={{
-                                    width: '100%',
-                                    color: Colors.GRAY1,
-                                }}
-                            >
-                                Tried to reference field with unknown id:{' '}
-                                {series.encode.yRef.field}
-                            </span>
-                        </SeriesBlock>
-                    );
-                }
-                return (
-                    <SeriesBlock>
-                        <SeriesTitle>{getItemLabel(field)}</SeriesTitle>
-                        <SeriesConfiguration
-                            series={series}
-                            placeholderName={getItemLabel(field)}
-                            fallbackColor={getSeriesColor(getSeriesId(series))}
-                            onColorChange={(color) => {
-                                updateSingleSeries({
-                                    ...series,
-                                    color,
-                                });
-                            }}
-                            onNameChange={(name) =>
-                                updateSingleSeries({
-                                    ...series,
-                                    name,
-                                })
-                            }
-                            onLabelChange={(label) => {
-                                updateSingleSeries({
-                                    ...series,
-                                    label,
-                                });
-                            }}
-                        />
-                    </SeriesBlock>
-                );
-            })}
-        </>
-    );
-};
+import BasicSeriesConfiguration from './Series/BasicSeriesConfiguration';
+import GroupedSeriesConfiguration from './Series/GroupedSeriesConfiguration';
 
 const ChartConfigTabs: FC = () => {
     const {
@@ -489,20 +201,7 @@ const ChartConfigTabs: FC = () => {
                         pivotDimension ? (
                             <GroupedSeriesConfiguration
                                 items={items}
-                                groupedSeries={
-                                    dirtyEchartsConfig?.series?.reduce<
-                                        Record<string, Series[]>
-                                    >(
-                                        (hash, obj) => ({
-                                            ...hash,
-                                            [obj.encode.yRef.field]: (
-                                                hash[obj.encode.yRef.field] ||
-                                                []
-                                            ).concat(obj),
-                                        }),
-                                        {},
-                                    ) || {}
-                                }
+                                series={dirtyEchartsConfig?.series}
                                 getSeriesColor={getSeriesColor}
                                 updateSingleSeries={updateSingleSeries}
                                 updateAllGroupedSeries={updateAllGroupedSeries}
